Reactors AI providers to simplify provider APIs - Splits code into separate folders, e.g. models, utils - Adds structured prompt template handling - Simplifies provider APIs and request handling
