+++ src/views/nodes/resultsCommitsNode.ts
@@ -1,10 +1,10 @@
 import { TreeItem, TreeItemCollapsibleState } from 'vscode';
-import type { AIGenerateChangelogChange } from '../../ai/aiProviderService';
 import { GitUri } from '../../git/gitUri';
 import { isStash } from '../../git/models/commit';
 import type { GitRevisionRange } from '../../git/models/revision';
 import type { CommitsQueryResults, FilesQueryResults } from '../../git/queryResults';
 import { getChangesForChangelog } from '../../git/utils/-webview/log.utils';
+import type { AIGenerateChangelogChange } from '../../plus/ai/aiProviderService';
 import { configuration } from '../../system/-webview/configuration';
 import { gate } from '../../system/decorators/-webview/gate';
 import { debug } from '../../system/decorators/log';
